+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+Removed
+^^^^^^^
+* Dropped python 3.7 support
+
 `0.4.0`_ - 2022-01-10
 ---------------------