+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+`0.5.0`_ - 2023-04-28
+---------------------
+
 Added
 ^^^^^
 * Added method to return all the pythagorean triplets
@@ -94,8 +97,9 @@ Added
 * Added a method to check whether a number is prime
 
 
-.. _`unreleased`: https://github.com/spapanik/mathlib/compare/v0.4.0...master
-.. _`0.4.0`: https://github.com/spapanik/mathlib/compare/v0.2.0...v0.4.0
+.. _`unreleased`: https://github.com/spapanik/mathlib/compare/v0.5.0...master
+.. _`0.5.0`: https://github.com/spapanik/mathlib/compare/v0.4.0...v0.5.0
+.. _`0.4.0`: https://github.com/spapanik/mathlib/compare/v0.3.0...v0.4.0
 .. _`0.3.0`: https://github.com/spapanik/mathlib/compare/v0.2.0...v0.3.0
 .. _`0.2.0`: https://github.com/spapanik/mathlib/compare/v0.1.4...v0.2.0
 .. _`0.1.4`: https://github.com/spapanik/mathlib/compare/v0.1.3...v0.1.4